Make test cases for spec generation
